refactor(nodes): move vector DB tool interface into store.py per review feedback
Rod-Christensen suggested co-locating the tool definitions with the
underlying store so every vectordb node can "morph into a tool" without
a separate generic wrapper node.

The engine dispatches ``invoke`` against the node's ``IInstance`` (not
the ``DocumentStoreBase`` subclass, which is composed via
``self.IGlobal.store``), so a mixin that lives next to
``DocumentStoreBase`` in ``store.py`` and is inherited by the vectordb
``IInstance`` classes achieves the spirit of the suggestion while being
technically correct.

Changes:

- Add ``VectorStoreToolMixin`` to ``packages/ai/src/ai/common/store.py``
  with ``@tool_function`` decorated ``search``, ``upsert`` and
  ``delete`` methods that delegate to ``self.IGlobal.store``. The mixin
  is picked up automatically by ``IInstanceBase._collect_tool_methods``
  via MRO walk.
- Have ``pinecone``, ``chroma`` and ``qdrant`` ``IInstance`` classes
  inherit from ``VectorStoreToolMixin`` so each vectordb node exposes
  the search/upsert/delete tools natively.
- Add ``"tool"`` to ``classType`` and ``"invoke"`` to ``capabilities``
  in each node's ``services.json`` so ``getControllerNodeIds('tool')``
  discovers them as tool nodes.
- Delete the ``tool_vectordb`` node entirely — obsolete now that every
  vectordb node has native tool capability.
- Delete the unused ``vectordb.svg`` icon that existed only for the
  now-deleted ``tool_vectordb`` node.
- Fix the wrapped-payload normalization bug flagged in CodeRabbit
  comment 3049766102: ``{'input': '{"query": "x"}'}`` and
  ``{'input': SomeModel(...)}`` now recurse through
  ``_normalize_vectordb_tool_input`` on the inner value, so string-,
  model- and dict-wrapped payloads all receive the same
  JSON-parse/``model_dump`` treatment before merging outer extras.

Other vectordb connectors (astra_db, milvus, weaviate, atlas,
vectordb_postgres, index_search) can opt in to the same tool interface
with a one-line change to their ``IInstance.py`` plus the services.json
update — out of scope for this PR.
