--- nodes/src/nodes/tool_vectordb/services.json
@@ -1,113 +1,0 @@
-{
-	"title": "Vector DB Tool",
-	"protocol": "tool_vectordb://",
-	"classType": ["tool"],
-	"capabilities": ["invoke"],
-	"register": "filter",
-	"node": "python",
-	"path": "nodes.tool_vectordb",
-	"prefix": "vectordb",
-	"icon": "vectordb.svg",
-	"description": [
-		"A generic tool interface that exposes vector database operations (search, upsert, delete) ",
-		"to LLM agents. Works with any configured vector DB backend — Pinecone, ChromaDB, or Qdrant ",
-		"— through a unified tool abstraction."
-	],
-	"tile": ["Backend: ${parameters.tool_vectordb.backend}", "Tool: ${parameters.tool_vectordb.serverName}.vectordb"],
-	"lanes": {},
-	"preconfig": {
-		"default": "default",
-		"profiles": {
-			"default": {
-				"title": "Default",
-				"serverName": "vectordb",
-				"backend": "pinecone"
-			}
-		}
-	},
-	"fields": {
-		"tool_vectordb.serverName": {
-			"type": "string",
-			"title": "Server name",
-			"description": "Namespace prefix for the tools: <serverName>.search, <serverName>.upsert, <serverName>.delete",
-			"default": "vectordb"
-		},
-		"tool_vectordb.backend": {
-			"title": "Vector DB Backend",
-			"description": "Which vector database backend to use",
-			"type": "string",
-			"default": "pinecone",
-			"enum": [
-				["pinecone", "Pinecone"],
-				["chroma", "ChromaDB"],
-				["qdrant", "Qdrant"]
-			]
-		},
-		"tool_vectordb.collection_description": {
-			"type": "string",
-			"title": "Collection description",
-			"default": "",
-			"optional": true,
-			"description": "Describe what is stored in this collection. This helps the agent search more accurately.",
-			"ui": {
-				"ui:widget": "textarea"
-			}
-		},
-		"tool_vectordb.enableSearch": {
-			"type": "boolean",
-			"title": "Enable search tool",
-			"default": true,
-			"enum": [
-				[true, "Yes"],
-				[false, "No"]
-			]
-		},
-		"tool_vectordb.enableUpsert": {
-			"type": "boolean",
-			"title": "Enable upsert tool",
-			"default": false,
-			"enum": [
-				[true, "Yes"],
-				[false, "No"]
-			]
-		},
-		"tool_vectordb.enableDelete": {
-			"type": "boolean",
-			"title": "Enable delete tool",
-			"default": false,
-			"enum": [
-				[true, "Yes"],
-				[false, "No"]
-			]
-		},
-		"tool_vectordb.topK": {
-			"type": "integer",
-			"title": "Default top_k",
-			"description": "Default number of results to return from search (agent can override)",
-			"default": 10
-		},
-		"tool_vectordb.scoreThreshold": {
-			"type": "number",
-			"title": "Score threshold",
-			"description": "Minimum similarity score to include in results (0.0 to 1.0)",
-			"default": 0.0
-		}
-	},
-	"shape": [
-		{
-			"section": "Pipe",
-			"title": "Vector DB Tool",
-			"properties": [
-				"type",
-				"tool_vectordb.serverName",
-				"tool_vectordb.backend",
-				"tool_vectordb.collection_description",
-				"tool_vectordb.enableSearch",
-				"tool_vectordb.enableUpsert",
-				"tool_vectordb.enableDelete",
-				"tool_vectordb.topK",
-				"tool_vectordb.scoreThreshold"
-			]
-		}
-	]
-}
